% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigatoR3-package.R
\docType{package}
\name{investigatoR3-package}
\alias{investigatoR3}
\alias{investigatoR3-package}
\title{investigatoR3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ericschumann12/investigatoR3}
  \item Report bugs at \url{https://github.com/ericschumann12/investigatoR3/issues}
}

}
\author{
\strong{Maintainer}: Eric Schumann \email{eric.schumann@uni.li}

}
\keyword{internal}
