% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Portfolio Performance Evaluation}
\usage{
performance(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
numbers
}
\description{
Portfolio Performance Evaluation
}
\examples{
performance(c(1,2,3,NA,5))

}
